ChangePassword service is tested and added. User model is transferred from authentication to common folder.
