+++ source/backend/api/common/data_check.py
@@ -1,7 +1,5 @@
 def check_data_keys(data: dict, necessary_keys: dict) -> str:
-    for key in data.keys():
-        if key not in necessary_keys:
-            return f'Redundant key {key}'
+    # will be fixed
 
     for key in necessary_keys:
         if key not in data.keys():
